Notas Fiscais/requirements.txt
streamlit
pandas
openai
langchain
langchain-openai
python-dotenv
langchain-google-genai